feat: add executor config creation draft

+++ common/src/api/probes.rs
@@ -5,5 +5,5 @@ pub struct Probes {
     #[serde(rename = "type")]
     pub probe_type: String, // should always be "probes"
     pub value: String,      // A comma seperated list of probe ids
-    pub requested: u32,     // The amount of comma seperated probe ids
+    pub requested: usize,     // The amount of comma seperated probe ids
 }
